Flyttet duplisert kode for beregning av trygdeavgiftsperioder til utils, fjernet grunnlagsopplysningerSkjema.tsx da den virket unødvendig
